+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixed an issue where the left sidebar arrow button was not always visible in exported projects
 - Removed Auto updater from MacOS
 - Faster loading times for languages
+- Cpp language highlighting for template functions
 
 ### Technical Changes
 - Change `state`, `PresetProject`, `PresetTheme` and `RecentProject` to use the `wrapEntity / unwrapEntity` structure
@@ -29,6 +30,7 @@
 - Added new migration files `PresetProjectMigration.js`, `PresetThemeMigration.js` and `RecentProjectMigration.js`
 - Added `TempFileManager` to detect and remove stale temporary files left behind by failed or interrupted write operations
 - Added dompurify package
+- Added balanced lookahead to syntax definition match
 
 <!-- update-meta: minCompatibleVersion="2.0.0"; incompatibilityNote="This version is not compatible with previous versions. Every project that was created needs to be exported as a .dfproj file to avoid being lost. The exported project can be imported into the new version without any loss."; -->